Converts all of power to use doubles so it doesn't lose precision past 16MW. Except for that Pow3r UI because IMGUI wants to torture me. @PJB3005!

+++ Content.Client/Power/APC/UI/ApcMenu.xaml.cs
@@ -55,8 +55,8 @@ namespace Content.Client.Power.APC.UI
 
             if (ChargeBar != null)
             {
-                ChargeBar.Value = castState.Charge;
-                UpdateChargeBarColor(castState.Charge);
+                ChargeBar.Value = (float)castState.Charge;
+                UpdateChargeBarColor((float)castState.Charge);
 
                 if (APCMenu != null)
                 {
