+++ Content.Server/Power/NodeGroups/ApcNet.cs
@@ -122,8 +122,8 @@ namespace Content.Server.Power.NodeGroups
 
             var ps = _powerNetSystem.GetNetworkStatistics(NetworkNode);
 
-            float storageRatio = ps.InStorageCurrent / Math.Max(ps.InStorageMax, 1.0f);
-            float outStorageRatio = ps.OutStorageCurrent / Math.Max(ps.OutStorageMax, 1.0f);
+            double storageRatio = ps.InStorageCurrent / Math.Max(ps.InStorageMax, 1.0f);
+            double outStorageRatio = ps.OutStorageCurrent / Math.Max(ps.OutStorageMax, 1.0f);
             return @$"Current Supply: {ps.SupplyCurrent:G3}
 From Batteries: {ps.SupplyBatteries:G3}
 Theoretical Supply: {ps.SupplyTheoretical:G3}
